+++ library/src/main/java/cloud/pace/sdk/poikit/poi/download/CmsAPI.kt
@@ -1,14 +1,17 @@
 package cloud.pace.sdk.poikit.poi.download
 
 import cloud.pace.sdk.api.API
+import cloud.pace.sdk.api.utils.EnumConverterFactory
 import cloud.pace.sdk.api.utils.InterceptorUtils
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.GET
+import java.util.*
 
 object CmsAPI {
     interface GetPaymentMethodVendorsService {
@@ -17,12 +20,18 @@ object CmsAPI {
     }
 
     private val service: GetPaymentMethodVendorsService by lazy {
+        val client = OkHttpClient.Builder()
+            .addNetworkInterceptor(InterceptorUtils.getInterceptor("application/json", "application/json", true))
+            .authenticator(InterceptorUtils.getAuthenticator())
+
         Retrofit.Builder()
-            .client(OkHttpClient.Builder().addInterceptor(InterceptorUtils.getInterceptor("application/json", "application/json", true)).build())
+            .client(client.build())
             .baseUrl(API.baseUrl)
+            .addConverterFactory(EnumConverterFactory())
             .addConverterFactory(
                 MoshiConverterFactory.create(
                     Moshi.Builder()
+                        .add(Date::class.java, Rfc3339DateJsonAdapter().nullSafe())
                         .add(KotlinJsonAdapterFactory())
                         .build()
                 )