+++ library/src/main/java/cloud/pace/sdk/poikit/pricehistory/PriceHistoryFuelType.kt
@@ -1,0 +1,9 @@
+package cloud.pace.sdk.poikit.pricehistory
+
+import java.util.*
+
+data class PriceHistoryFuelType(
+    val time: Date,
+    val price: Double,
+    val currency: String
+)